move scenarios into its own file
